CHANGE: refactoring: - renamed class modulest by CamelCase convention; - renamed class to MatchInfoCombo.
